finished

+++ 02_printf/libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: mmatsego <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/13 09:41:32 by mmatsego          #+#    #+#             */
-/*   Updated: 2021/02/05 16:54:58 by mmatsego         ###   ########.fr       */
+/*   Updated: 2021/02/08 14:49:34 by mmatsego         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,7 +30,7 @@ int					ft_isalpha(int c);
 int					ft_isascii(int c);
 int					ft_isdigit(int c);
 int					ft_isprint(int c);
-char				*ft_itoa(int n);
+char				*ft_itoa(size_t n);
 char				*ft_itoa_base(unsigned long nbr, int base, char c);
 void				ft_lstadd_back(t_list **lst, t_list *new);
 void				ft_lstadd_front(t_list **lst, t_list *new);
